FittingWorker -> FitWorker (files renaming)
