+++ GUI/coregui/Views/FitWidgets/FitWorker.h
